chore: 배너 디테일 페이지 라우팅 및 폴더 구조 정리
