+++ src/pages/banner/BannerDetailPage.css.ts
